# bpo-45636: Simpler design for `BINARY_OP`/`INPLACE_OP`