… Doc/library/dis.rst
… Doc/whatsnew/3.11.rst
… Include/opcode.h
… Lib/test/test_dis.py
… Misc/NEWS.d/next/Core and Builtins/2021-10-27-15-14-31.bpo-45636.K2X7QS.rst
… Python/opcode_targets.h