No perf impact:

<details>

| Benchmark               | main    | binary-op-alt         |
|-------------------------|:-------:|:---------------------:|
| nbody                   | 118 ms  | 113 ms: 1.05x faster  |
| logging_silent          | 118 ns  | 114 ns: 1.04x faster  |
| xml_etree_iterparse     | 109 ms  | 106 ms: 1.03x faster  |
| go                      | 172 ms  | 167 ms: 1.02x faster  |
| richards                | 57.8 ms | 56.8 ms: 1.02x faster |
| dulwich_log             | 67.7 ms | 66.7 ms: 1.01x faster |
| telco                   | 6.33 ms | 6.24 ms: 1.01x faster |
| pickle_pure_python      | 369 us  | 364 us: 1.01x faster  |
| scimark_sparse_mat_mult | 4.74 ms | 4.69 ms: 1.01x faster |
| hexiom                  | 7.53 ms | 7.44 ms: 1.01x faster |
| meteor_contest          | 107 ms  | 106 ms: 1.01x faster  |
| xml_etree_parse         | 155 ms  | 154 ms: 1.01x faster  |
| pickle_list             | 4.58 us | 4.56 us: 1.01x faster |
| regex_compile           | 145 ms  | 144 ms: 1.00x faster  |
| nqueens                 | 88.5 ms | 88.1 ms: 1.00x faster |
| sympy_sum               | 170 ms  | 169 ms: 1.00x faster  |
| python_startup          | 8.53 ms | 8.52 ms: 1.00x faster |
| python_startup_no_site  | 5.90 ms | 5.91 ms: 1.00x slower |
| pickle                  | 9.90 us | 9.94 us: 1.00x slower |
| pidigits                | 194 ms  | 195 ms: 1.00x slower  |
| fannkuch                | 417 ms  | 419 ms: 1.01x slower  |
| json_dumps              | 12.5 ms | 12.5 ms: 1.01x slower |
| chaos                   | 76.5 ms | 76.9 ms: 1.01x slower |
| unpickle                | 13.4 us | 13.5 us: 1.01x slower |
| pyflate                 | 534 ms  | 538 ms: 1.01x slower  |
| mako                    | 11.9 ms | 12.0 ms: 1.01x slower |
| json_loads              | 25.4 us | 25.7 us: 1.01x slower |
| logging_format          | 6.73 us | 6.79 us: 1.01x slower |
| logging_simple          | 6.06 us | 6.13 us: 1.01x slower |
| chameleon               | 7.53 ms | 7.62 ms: 1.01x slower |
| scimark_fft             | 347 ms  | 352 ms: 1.01x slower  |
| regex_dna               | 216 ms  | 219 ms: 1.01x slower  |
| pathlib                 | 18.7 ms | 19.0 ms: 1.02x slower |
| unpickle_list           | 4.95 us | 5.05 us: 1.02x slower |
| spectral_norm           | 107 ms  | 110 ms: 1.03x slower  |
| scimark_monte_carlo     | 78.0 ms | 80.4 ms: 1.03x slower |
| regex_v8                | 23.2 ms | 24.2 ms: 1.04x slower |
| pickle_dict             | 27.5 us | 28.7 us: 1.04x slower |
| regex_effbot            | 3.18 ms | 3.34 ms: 1.05x slower |
| unpack_sequence         | 44.0 ns | 48.8 ns: 1.11x slower |
| Geometric mean          | (ref)   | 1.00x slower          |

</details>

<!-- issue-number: [bpo-45636](https://bugs.python.org/issue45636) -->
https://bugs.python.org/issue45636
<!-- /issue-number -->