@brandtbucher 2021-11-05
I'm glad I did. It seems that [that single change](https://github.com/python/cpython/pull/29418/commits/bd8e9c711664f2871c120a52e60965d9c8e40a83) contributes to the slowdown we saw on the other PR:

<details>

```
Slower (37):
- crypto_pyaes: 87.5 ms +- 0.6 ms -> 91.3 ms +- 1.1 ms: 1.04x slower
- pickle_dict: 26.8 us +- 0.1 us -> 27.9 us +- 4.1 us: 1.04x slower
- go: 165 ms +- 1 ms -> 171 ms +- 2 ms: 1.04x slower
- richards: 56.9 ms +- 0.9 ms -> 58.8 ms +- 0.7 ms: 1.03x slower
- chameleon: 7.48 ms +- 0.10 ms -> 7.71 ms +- 0.07 ms: 1.03x slower
- regex_compile: 144 ms +- 2 ms -> 148 ms +- 1 ms: 1.03x slower
- spectral_norm: 107 ms +- 1 ms -> 110 ms +- 2 ms: 1.02x slower
- tornado_http: 109 ms +- 3 ms -> 111 ms +- 3 ms: 1.02x slower
- scimark_monte_carlo: 77.5 ms +- 0.8 ms -> 79.1 ms +- 1.4 ms: 1.02x slower
- chaos: 76.7 ms +- 0.7 ms -> 78.3 ms +- 2.5 ms: 1.02x slower
- deltablue: 4.85 ms +- 0.06 ms -> 4.95 ms +- 0.07 ms: 1.02x slower
- nbody: 115 ms +- 2 ms -> 118 ms +- 3 ms: 1.02x slower
- sympy_expand: 501 ms +- 5 ms -> 510 ms +- 5 ms: 1.02x slower
- float: 80.5 ms +- 0.9 ms -> 81.8 ms +- 0.9 ms: 1.02x slower
- json_loads: 25.4 us +- 0.2 us -> 25.8 us +- 0.2 us: 1.02x slower
- pickle_pure_python: 367 us +- 3 us -> 373 us +- 4 us: 1.02x slower
- regex_dna: 212 ms +- 1 ms -> 215 ms +- 2 ms: 1.01x slower
- pyflate: 539 ms +- 9 ms -> 545 ms +- 4 ms: 1.01x slower
- unpickle_pure_python: 272 us +- 3 us -> 275 us +- 3 us: 1.01x slower
- sqlite_synth: 2.74 us +- 0.04 us -> 2.77 us +- 0.07 us: 1.01x slower
- sympy_integrate: 22.0 ms +- 0.2 ms -> 22.2 ms +- 0.1 ms: 1.01x slower
- django_template: 37.2 ms +- 0.5 ms -> 37.6 ms +- 0.4 ms: 1.01x slower
- scimark_sparse_mat_mult: 4.84 ms +- 0.08 ms -> 4.89 ms +- 0.14 ms: 1.01x slower
- xml_etree_iterparse: 107 ms +- 3 ms -> 108 ms +- 3 ms: 1.01x slower
- pickle_list: 4.48 us +- 0.04 us -> 4.52 us +- 0.03 us: 1.01x slower
- sympy_str: 303 ms +- 4 ms -> 306 ms +- 3 ms: 1.01x slower
- meteor_contest: 106 ms +- 2 ms -> 106 ms +- 2 ms: 1.01x slower
- unpickle_list: 5.05 us +- 0.04 us -> 5.08 us +- 0.07 us: 1.01x slower
- xml_etree_process: 59.1 ms +- 0.6 ms -> 59.5 ms +- 0.4 ms: 1.01x slower
- xml_etree_generate: 80.4 ms +- 0.7 ms -> 80.9 ms +- 1.0 ms: 1.01x slower
- 2to3: 271 ms +- 1 ms -> 273 ms +- 1 ms: 1.01x slower
- hexiom: 7.53 ms +- 0.12 ms -> 7.57 ms +- 0.07 ms: 1.01x slower
- sympy_sum: 170 ms +- 2 ms -> 171 ms +- 2 ms: 1.01x slower
- dulwich_log: 67.6 ms +- 0.3 ms -> 67.9 ms +- 0.4 ms: 1.00x slower
- regex_v8: 23.2 ms +- 0.1 ms -> 23.3 ms +- 0.2 ms: 1.00x slower
- python_startup: 8.51 ms +- 0.01 ms -> 8.54 ms +- 0.01 ms: 1.00x slower
- python_startup_no_site: 5.90 ms +- 0.00 ms -> 5.91 ms +- 0.00 ms: 1.00x slower

Faster (6):
- scimark_fft: 355 ms +- 10 ms -> 346 ms +- 2 ms: 1.03x faster
- fannkuch: 427 ms +- 4 ms -> 417 ms +- 2 ms: 1.02x faster
- pidigits: 190 ms +- 0 ms -> 187 ms +- 0 ms: 1.01x faster
- json_dumps: 12.6 ms +- 0.1 ms -> 12.5 ms +- 0.1 ms: 1.01x faster
- pathlib: 19.0 ms +- 0.3 ms -> 18.8 ms +- 0.2 ms: 1.01x faster
- mako: 11.9 ms +- 0.1 ms -> 11.9 ms +- 0.1 ms: 1.00x faster

Benchmark hidden because not significant (15): logging_format, logging_silent, logging_simple, nqueens, pickle, raytrace, regex_effbot, scimark_lu, scimark_sor, sqlalchemy_declarative, sqlalchemy_imperative, telco, unpack_sequence, unpickle, xml_etree_parse

Geometric mean: 1.01x slower
```

</details>

It's not clear to me why. Perhaps `int` is that much faster than `uint16_t`?