@brandtbucher 2021-11-09
You were right @markshannon: adjusting the alignment fixes the slowdown. Nice catch.

<details>

```
Slower (25):
- pickle_dict: 26.8 us +- 0.1 us -> 28.0 us +- 0.1 us: 1.04x slower
- go: 165 ms +- 1 ms -> 172 ms +- 2 ms: 1.04x slower
- regex_effbot: 3.15 ms +- 0.03 ms -> 3.25 ms +- 0.04 ms: 1.03x slower
- regex_dna: 212 ms +- 1 ms -> 219 ms +- 1 ms: 1.03x slower
- scimark_lu: 138 ms +- 3 ms -> 143 ms +- 5 ms: 1.03x slower
- float: 80.5 ms +- 0.9 ms -> 82.8 ms +- 1.1 ms: 1.03x slower
- logging_format: 6.73 us +- 0.09 us -> 6.89 us +- 0.08 us: 1.02x slower
- chameleon: 7.48 ms +- 0.10 ms -> 7.64 ms +- 0.09 ms: 1.02x slower
- regex_v8: 23.2 ms +- 0.1 ms -> 23.7 ms +- 0.4 ms: 1.02x slower
- tornado_http: 109 ms +- 3 ms -> 111 ms +- 3 ms: 1.02x slower
- scimark_monte_carlo: 77.5 ms +- 0.8 ms -> 79.0 ms +- 1.3 ms: 1.02x slower
- nqueens: 90.1 ms +- 0.7 ms -> 91.3 ms +- 1.0 ms: 1.01x slower
- django_template: 37.2 ms +- 0.5 ms -> 37.7 ms +- 0.6 ms: 1.01x slower
- sympy_expand: 501 ms +- 5 ms -> 507 ms +- 3 ms: 1.01x slower
- sqlalchemy_imperative: 18.6 ms +- 0.5 ms -> 18.8 ms +- 0.6 ms: 1.01x slower
- richards: 56.9 ms +- 0.9 ms -> 57.6 ms +- 1.4 ms: 1.01x slower
- regex_compile: 144 ms +- 2 ms -> 146 ms +- 2 ms: 1.01x slower
- unpack_sequence: 43.7 ns +- 0.9 ns -> 44.2 ns +- 1.1 ns: 1.01x slower
- xml_etree_generate: 80.4 ms +- 0.7 ms -> 81.2 ms +- 0.8 ms: 1.01x slower
- json_loads: 25.4 us +- 0.2 us -> 25.6 us +- 0.2 us: 1.01x slower
- mako: 11.9 ms +- 0.1 ms -> 12.0 ms +- 0.0 ms: 1.01x slower
- pickle_list: 4.48 us +- 0.04 us -> 4.52 us +- 0.04 us: 1.01x slower
- scimark_sparse_mat_mult: 4.84 ms +- 0.08 ms -> 4.88 ms +- 0.05 ms: 1.01x slower
- logging_simple: 6.15 us +- 0.08 us -> 6.19 us +- 0.07 us: 1.01x slower
- unpickle_list: 5.05 us +- 0.04 us -> 5.07 us +- 0.05 us: 1.00x slower

Faster (15):
- scimark_fft: 355 ms +- 10 ms -> 345 ms +- 2 ms: 1.03x faster
- fannkuch: 427 ms +- 4 ms -> 419 ms +- 4 ms: 1.02x faster
- xml_etree_iterparse: 107 ms +- 3 ms -> 105 ms +- 2 ms: 1.02x faster
- telco: 6.38 ms +- 0.25 ms -> 6.28 ms +- 0.11 ms: 1.02x faster
- pickle: 9.97 us +- 0.16 us -> 9.84 us +- 0.11 us: 1.01x faster
- pidigits: 190 ms +- 0 ms -> 188 ms +- 0 ms: 1.01x faster
- xml_etree_parse: 155 ms +- 3 ms -> 153 ms +- 1 ms: 1.01x faster
- crypto_pyaes: 87.5 ms +- 0.6 ms -> 86.6 ms +- 1.1 ms: 1.01x faster
- hexiom: 7.53 ms +- 0.12 ms -> 7.47 ms +- 0.06 ms: 1.01x faster
- raytrace: 331 ms +- 2 ms -> 329 ms +- 1 ms: 1.01x faster
- json_dumps: 12.6 ms +- 0.1 ms -> 12.5 ms +- 0.1 ms: 1.01x faster
- spectral_norm: 107 ms +- 1 ms -> 107 ms +- 1 ms: 1.01x faster
- python_startup: 8.51 ms +- 0.01 ms -> 8.49 ms +- 0.01 ms: 1.00x faster
- dulwich_log: 67.6 ms +- 0.3 ms -> 67.4 ms +- 0.6 ms: 1.00x faster
- python_startup_no_site: 5.90 ms +- 0.00 ms -> 5.89 ms +- 0.00 ms: 1.00x faster

Benchmark hidden because not significant (18): 2to3, chaos, deltablue, logging_silent, meteor_contest, nbody, pathlib, pickle_pure_python, pyflate, scimark_sor, sqlalchemy_declarative, sqlite_synth, sympy_integrate, sympy_sum, sympy_str, unpickle, unpickle_pure_python, xml_etree_process

Geometric mean: 1.00x slower
```

</details>